pyproject.toml
[project]
name = "bird-mcp"
version = "0.1.0"
description = "Personal assistant MCP server with Todoist, Anki, Obsidian, and Google Calendar integration"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "todoist-api-python>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
